move all _case_references tests to same module
